refactor: extract shared resource docs helpers
