+++ .content-fallback-allowlist
@@ -27,12 +27,10 @@ docs/claude-skills/
 # Kept as defense-in-depth, not because the route is unprotected (#3259
 # supersedes the earlier "most exposed due to size" framing — do not
 # reintroduce it). Two structural defenses already guard generated skill
-# pages: `escapeForMdx()` runs over every generated body
-# (packages/zudo-doc/src/plugins/internal/claude-resources/generate.ts:580,
-# subpages :610/:625/:640) and escapes braces and non-HTML tags to entities
-# (escape-for-mdx.ts:53-86), and the output dir is wiped via a recursive
-# `rmSync` and regenerated every build (`cleanDir` at generate.ts:497, helper
-# at :65-68), so hand-placed raw MDX never survives. The entry stays because
+# pages: the shared `escapeForMdx` helper runs over every generated body and
+# subpage and escapes braces and non-HTML tags to entities, and the output dir
+# is wiped via the shared `cleanDir` helper and regenerated every build, so
+# hand-placed raw MDX never survives. The entry stays because
 # the upstream trigger (Takazudo/zudo-front-builder#2186, epic #3232) is
 # shape-sensitive and size-driven (single concatenated dumps —
 # l-lessons-zfb-migration-parity is ~150 KB) rather than tied to a specific
